ignore -index-store-path option
This option is used to enable 'Indexing while Building' in Xcode9,
which does not affect the build output.

The reason we have to skip it is it introduces a path contains
uuid, which causes build command is different on each
machines. So it's better to skip it in manifest hashing.